LUCENE-5767: remove bogus cast (in this case can exceed Integer.MAX_VALUE, and the underlying delta reader takes long anyway)
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1604064 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/index/MultiDocValues.java
@@ -484,7 +484,7 @@ public class MultiDocValues {
             segmentToGlobalOrds[i] = new LongValues() {
               @Override
               public long get(long ord) {
-                return ord + deltas.get((int) ord);
+                return ord + deltas.get(ord);
               }
             };
             ramBytesUsed += deltas.ramBytesUsed();
